# Milky Way 2.0/retail_app/docs/developer/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Milky Way 2.0 · Retail Agent",
    "version": "2.1.1"
  },
  "paths": {
...
    "/api/evaluations/runs": {
      "get": {
        "tags": [
          "Workspace evaluations"
        ],
        "summary": "Runs",
        "operationId": "runs_api_evaluations_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workspace evaluations"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_evaluations_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations/runs/{run_id}": {
      "get": {
        "tags": [
          "Workspace evaluations"
        ],
        "summary": "Run",
        "operationId": "run_api_evaluations_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "Workspace evaluations"
        ],
        "summary": "Cancel",
        "operationId": "cancel_api_evaluations_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations/runs/{run_id}/review": {
      "post": {
        "tags": [
          "Workspace evaluations"
        ],
        "summary": "Review",
        "operationId": "review_api_evaluations_runs__run_id__review_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewRequest": {
        "properties": {
          "reviewer": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Reviewer"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "notes": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reviewer",
          "decision",
          "notes"
        ],
        "title": "ReviewRequest"
      },
      "RunRequest": {
        "properties": {
          "suite_id": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Suite Id"
          },
          "baseline_release_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Release Id"
          },
          "candidate_release_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Release Id"
          },
          "candidate_version_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Version Ids"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "deterministic"
          },
          "max_cases": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Max Cases",
            "default": 120
          },
          "max_model_calls": {
            "type": "integer",
            "maximum": 40.0,
            "minimum": 0.0,
            "title": "Max Model Calls",
            "default": 0
          },
          "max_tokens": {
            "type": "integer",
            "maximum": 200000.0,
            "minimum": 0.0,
            "title": "Max Tokens",
            "default": 0
          },
          "confirm_billable": {
            "type": "boolean",
            "title": "Confirm Billable",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "suite_id"
        ],
        "title": "RunRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}